# web/src/lib/openapi.schema.json
{
  "components": {
    "schemas": {
...
      "EnvSetRequest": {
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "persist": {
            "default": true,
            "title": "Persist",
            "type": "boolean"
          },
          "value": {
            "format": "password",
            "title": "Value",
            "type": "string",
            "writeOnly": true
          }
        },
        "required": [
          "name",
          "value"
        ],
        "title": "EnvSetRequest",
        "type": "object"
      },
      "EnvSetResponse": {
        "properties": {
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "persisted": {
            "default": false,
            "title": "Persisted",
            "type": "boolean"
          },
          "session_only": {
            "default": true,
            "title": "Session Only",
            "type": "boolean"
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "success",
          "name"
        ],
        "title": "EnvSetResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "API server for RDST web client",
    "title": "RDST API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/env/set": {
      "post": {
        "operationId": "set_env_secret_api_env_set_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnvSetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvSetResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Env Secret",
        "tags": [
          "env"
        ]
      }
    },
...
  }
}